requirements.txt
pydantic
fastapi
uvicorn[standard]
motor
jinja2
python-dotenv
python-multipart